requirements.txt
fastapi==0.85.1
pytest==7.1.3
mysql-connector-python==8.0.31
numpy==1.21.4
mariadb==1.1.5.post2